==> RMD Files/2 Merging Geographical and HHI Data; 4 Visualizations.Rmd <==
---
title: "HHI PH"
author: "Rafael Ma. Vicente Tagulao"
date: "2024-12-05"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(sf)
library(dplyr)
library(stringdist)
library(tidyr)
library(ggplot2)
```


```{r}
#Reading the calculated HHI values
hhi = read.csv("datasets/unweighted_hhi_yearly.csv")
```

```{r}
# Reading shapefile obtained from GitHub
ph <- st_read("datasets/raw_geo_dataset/Provdists.shp.shp")

# Merging Maguindanao del Norte and Maguindanao del Sur into Maguindanao to fit HHI values
ph <- ph %>%
  mutate(name = ifelse(name %in% c("Maguindanao del Norte", "Maguindanao del Sur"),
                       "Maguindanao", name)) %>%
  group_by(name) %>%
  summarize(geometry = st_combine(geometry), .groups = "drop")

# Ensuring that the geometry has no intersecting areas
ph <- ph %>%
  mutate(geometry = st_make_valid(geometry))

# Calculating the centroid of polygons and extracting the latitude and longitude
ph <- ph %>%
  mutate(
    centroid = st_centroid(geometry), 
    lon = st_coordinates(centroid)[, 1],  
    lat = st_coordinates(centroid)[, 2]   
  )
```

```{r}
# Changing the columns to lower case, to match them better, and removing a null row in the shapefile
ph <- ph %>% mutate(name_lower = tolower(name))
hhi <- hhi %>% mutate(X_lower = tolower(X))
ph <- ph %>%
  filter(!is.na(name_lower))

# Manually assigning matching
hhi <- hhi %>%
  mutate(
    X_lower = case_when(
      X_lower == "ncr, city of manila, first district" ~ "ncr, city of manila, first district (not a province)",
      X_lower == "ncr, second district" ~ "ncr, second district (not a province)",
      X_lower == "ncr, third district" ~ "ncr, third district (not a province)",
      X_lower == "ncr, fourth district" ~ "ncr, fourth district (not a province)",
      X_lower == "compostela valley" ~ "davao de oro",
      TRUE ~ X_lower
    )
  )

hhi_map <- hhi %>%
  left_join(ph, by = c("X_lower" = "name_lower"))

#This is the the object we will then use for the visualization
hhi_map <- hhi_map %>%
  select(X, X2004, X2007, X2010, X2013, X2016, geometry, lon, lat)

# Exporting it now to shapefile
st_write(hhi_map, "hhi_map.shp", delete_dsn = TRUE)
```

```{r}
# Now, adding island group names

hhi_map <- hhi_map %>%
  mutate(
    Island_Group = case_when(
      X %in% c("ABRA", "ALBAY", "APAYAO", "AURORA", "BATAAN", "BATANES", "BATANGAS",
               "BENGUET", "BULACAN", "CAGAYAN", "CAMARINES NORTE", "CAMARINES SUR",
               "CATANDUANES", "CAVITE", "IFUGAO", "ILOCOS NORTE", "ILOCOS SUR",
               "ISABELA", "KALINGA", "LA UNION", "LAGUNA", "MARINDUQUE", "MASBATE",
               "MOUNTAIN PROVINCE", "NUEVA ECIJA", "NUEVA VIZCAYA", "OCCIDENTAL MINDORO",
               "ORIENTAL MINDORO", "PALAWAN", "PAMPANGA", "PANGASINAN", "QUEZON", 
               "QUIRINO", "RIZAL", "ROMBLON", "SORSOGON", "TARLAC", "ZAMBALES", 
               "NCR, SECOND DISTRICT", "NCR, CITY OF MANILA, FIRST DISTRICT", 
               "NCR, FOURTH DISTRICT", "NCR, THIRD DISTRICT") ~ "Luzon",
      X %in% c("AKLAN", "ANTIQUE", "BILIRAN", "BOHOL", "CAPIZ", "CEBU", "EASTERN SAMAR",
               "GUIMARAS", "ILOILO", "LEYTE", "NEGROS OCCIDENTAL", "NEGROS ORIENTAL",
               "NORTHERN SAMAR", "SAMAR", "SIQUIJOR", "SOUTHERN LEYTE") ~ "Visayas",
      X %in% c("AGUSAN DEL NORTE", "AGUSAN DEL SUR", "BASILAN", "BUKIDNON", 
               "CAMIGUIN", "COMPOSTELA VALLEY", "COTABATO", "DAVAO DE ORO", "DAVAO DEL NORTE", "DAVAO DEL SUR",
               "DAVAO OCCIDENTAL", "DAVAO ORIENTAL", "DINAGAT ISLANDS", "LANAO DEL NORTE",
               "LANAO DEL SUR", "MAGUINDANAO", 
               "MISAMIS OCCIDENTAL", "MISAMIS ORIENTAL", "SARANGANI", "SOUTH COTABATO", 
               "SULTAN KUDARAT", "SULU", "SURIGAO DEL NORTE", "SURIGAO DEL SUR", 
               "TAWI-TAWI", "ZAMBOANGA DEL NORTE", "ZAMBOANGA DEL SUR", 
               "ZAMBOANGA SIBUGAY") ~ "Mindanao",
      TRUE ~ NA_character_ 
    )
  )
```

```{r}
# Converting the dataframe into long
names(hhi_map)[-1] <- gsub("^X", "", names(hhi_map)[-1])
data_long <- hhi_map %>%
  pivot_longer(
    cols = c("2004", "2007", "2010", "2013", "2016"),  
    names_to = "Year",         
    values_to = "Value"        
  ) %>%
  mutate(
    Year = as.numeric(Year),   
    Island_Group = factor(Island_Group, levels = c("Luzon", "Visayas", "Mindanao"))  
  ) %>%
  filter(Value != 0) 
```

```{r}
# Box Plot
ggplot(data_long, aes(x = as.factor(Year), y = Value, fill = Island_Group)) +
  geom_boxplot() + 
  labs(title = "HHI Value by Island Group", 
       x = "Year", y = "HHI Value") +
  theme_minimal() +
  scale_fill_manual(
    values = c(
      "Luzon" = "#BA3A2C",       
      "Visayas" = "#437C90",     
      "Mindanao" = "#A98743"     
    )
  )
```

```{r}
# Computing for median of each group
median_data <- data_long %>%
  group_by(Island_Group, Year) %>%
  summarize(Median_Value = median(Value, na.rm = TRUE)) %>%
  ungroup()

# Computing  the national median
national_median <- data_long %>%
  group_by(Year) %>%
  summarize(Median_Value = median(Value, na.rm = TRUE)) %>%
  mutate(Island_Group = "National")

combined_data <- bind_rows(median_data, national_median)
```

```{r}
combined_data$Island_Group <- factor(combined_data$Island_Group, levels = c("Luzon", "Visayas", "Mindanao", "National"))

# LINE PLOT
ggplot(combined_data, aes(x = Year, y = Median_Value, color = Island_Group, group = Island_Group)) +
  geom_line(size = 1) +  
  geom_point(size = 2) +  
  labs(title = "Median HHI Value by Island Group and National", 
       x = "Year", y = "Median HHI Value") +
  scale_color_manual(
    values = c(
      "Luzon" = "#BA3A2C",       
      "Visayas" = "#437C90",     
      "Mindanao" = "#A98743",    
      "National" = "#222222"     
    )
  ) +
  theme_minimal()  
```

```{r}
# Colors for the island groups
cols <- c("Luzon" = "#F76D5E", "Visayas" = "#72D8FF", "Mindanao" = "#FFFFBF")

# DENSITY PLOT
ggplot(data_long, aes(x = Value, fill = Island_Group, color = Island_Group)) +
  geom_density(alpha = 0.2) +  
  scale_fill_manual(values = cols) +  
  scale_color_manual(values = cols) +  
  facet_wrap(~ Year) + 
  labs(title = "HHI Density by Island Group and Year", 
       x = "HHI Value", y = "Density") +
  theme_minimal() +
  theme(
    plot.background = element_rect(fill = "#222222"),  
    panel.background = element_rect(fill = "#222222"),  
    panel.grid.major = element_line(color = "#555555"),  
    panel.grid.minor = element_line(color = "#555555"),  
    axis.text = element_text(color = "white"),  
    axis.title = element_text(color = "white"), 
    plot.title = element_text(color = "white"),  
    legend.text = element_text(color = "white"),  
    legend.title = element_text(color = "white"),  
    strip.text = element_text(color = "white")  
  )
```

```{r}
# PLOTTING IN MAP
#Year 2004
hhi_map <- st_as_sf(hhi_map)

# Create bins for HHI values
hhi_map$hhi_bins <- cut(
  hhi_map$`2004`,
  breaks = c(0, 50, 100, 150, 200, 250, 300, Inf),  
  labels = c("0-50", "50-100", "100-150", "150-200", "200-250", "250-300", "300+"),  
  include.lowest = TRUE
)

category_colors <- c(
  "0-50" = "#FFF3E0",
  "50-100" = "#FFE0C7",
  "100-150" = "#FFB899",
  "150-200" = "#FF8F70",
  "200-250" = "#FF5F4C",
  "250-300" = "#FF2A2A",
  "300+" = "#D30000"
)

plot1 <- ggplot(hhi_map) +
  geom_sf(aes(fill = hhi_bins), color = "black", alpha = 1) +  
  scale_fill_manual(values = category_colors, na.value = "grey80", name = "HHI Categories") +  
  theme_void() +  
  ggtitle("HHI Values in 2004") + 
  theme(
    legend.position = "right",  
    plot.title = element_text(size = 12)  
  )

plot1
```

```{r}
#Year 2007
hhi_map$hhi_bins <- cut(
  hhi_map$`2007`,
  breaks = c(0, 50, 100, 150, 200, 250, 300, Inf),  
  labels = c("0-50", "50-100", "100-150", "150-200", "200-250", "250-300", "300+"),  
  include.lowest = TRUE
)

plot2 <- ggplot(hhi_map) +
  geom_sf(aes(fill = hhi_bins), color = "black", alpha = 1) +  
  scale_fill_manual(values = category_colors, na.value = "grey80", name = "HHI Categories") +  
  theme_void() + 
  ggtitle("HHI Values in 2007") + 
  theme(
    legend.position = "right",  
    plot.title = element_text(size = 12) 
  )

plot2
```

```{r}
#Year 2010
hhi_map$hhi_bins <- cut(
  hhi_map$`2010`,
  breaks = c(0, 50, 100, 150, 200, 250, 300, Inf),  
  labels = c("0-50", "50-100", "100-150", "150-200", "200-250", "250-300", "300+"),  
  include.lowest = TRUE
)

plot3 <- ggplot(hhi_map) +
  geom_sf(aes(fill = hhi_bins), color = "black", alpha = 1) +  
  scale_fill_manual(values = category_colors, na.value = "grey80", name = "HHI Categories") + 
  theme_void() + 
  ggtitle("HHI Values in 2010") +  
  theme(
    legend.position = "right", 
    plot.title = element_text(size = 12) 
  )

plot3
```

```{r}
#Year 2013
hhi_map$hhi_bins <- cut(
  hhi_map$`2013`,
  breaks = c(0, 50, 100, 150, 200, 250, 300, Inf),  
  labels = c("0-50", "50-100", "100-150", "150-200", "200-250", "250-300", "300+"),  
  include.lowest = TRUE
)

plot4 <- ggplot(hhi_map) +
  geom_sf(aes(fill = hhi_bins), color = "black", alpha = 1) +  
  scale_fill_manual(values = category_colors, na.value = "grey80", name = "HHI Categories") +  
  theme_void() +  
  ggtitle("HHI Values in 2013") +  
  theme(
    legend.position = "right",  
    plot.title = element_text(size = 12)  
  )

plot4
```

```{r}
#Year 2016
hhi_map$hhi_bins <- cut(
  hhi_map$`2016`,
  breaks = c(0, 50, 100, 150, 200, 250, 300, Inf),  
  labels = c("0-50", "50-100", "100-150", "150-200", "200-250", "250-300", "300+"), 
  include.lowest = TRUE
)

plot5 <- ggplot(hhi_map) +
  geom_sf(aes(fill = hhi_bins), color = "black", alpha = 1) +  
  scale_fill_manual(values = category_colors, na.value = "grey80", name = "HHI Categories") +  
  theme_void() + 
  ggtitle("HHI Values in 2016") +  
  theme(
    legend.position = "right",  
    plot.title = element_text(size = 12)  
  )

plot5
```
