==> RMD Files/3 FIES Data Cleaning.Rmd <==
---
title: "FIES Data Cleaning"
author: "Sted Micah Cheng"
date: "`r Sys.Date()`"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# libraries needed
library(dplyr)
library(sf)
```


Import data
```{r}
fies_2006 = read.csv('datasets/fies_datasets/FIES PUF 2006 Vol.1.csv')
fies_2009 = read.csv('datasets/fies_datasets/FIES PUF 2009 Vol.2.csv')
fies_2012 = read.csv('datasets/fies_datasets/FIES PUF 2012 Vol.1.csv')
fies_2015 = read.csv('datasets/fies_datasets/FIES PUF 2015 Vol.1.csv')
fies_2018 = read.csv('datasets/fies_datasets/FIES PUF 2018 Vol.1.csv')
hhi_map = st_read('datasets/hhi_map/hhi_map.shp')
```


Select the relevant column names
```{r}
col_names_2006 = c('W_HCN', 'FSIZE', 'WAGES', 'NETSH', 'CONAB', 'CONDO', 'RENTL', 'INTRS', 
                   'PNSNS', 'DVDND', 'OTHIN', 'IFAMS', 'REGFT', 'EACFG', 'EALPR', 'EAFIS', 
                   'EAFOR', 'EATRD', 'EAMFG', 'EACPS', 'EATCS', 'EAMNG', 'EACON', 'EANEC', 'EAINC', 
                   'TOINC', 'CREAL', 'MEAT', 'FISHM', 'DAIRY', 'FRUIT', 'FDNEC', 'COFCT', 'NONAL', 
                   'ALBEV', 'TBCCO', 'FHOME', 'FDOUT', 'FOOD', 'CLOTH', 'NDFUR', 'MEDIC', 
                   'FUEL', 'TRCOM', 'RCRTN', 'EDUC', 'OTHEX', 
                   'OTDIS', 'NONFOOD', 'TOEXP', 'TOTDIS', 'OTHRE', 'TOREC')

col_names_2009 = c('W_ID', 'FSIZE', 'WAGES', 'NETSH', 'CONAB', 'CONDO', 'RENTL', 'INTRS', 
                   'PNSNS', 'DVDND', 'OSINC', 'IFAMS', 'REGFT', 'EACFG', 'EALPR', 'EAFIS', 
                   'EAFOR', 'EATRD', 'EAMFG', 'EACPS', 'EATCS', 'EAMNG', 'EACON', 'EANEC', 'EAINC', 
                   'TOINC', 'CREAL', 'MEAT', 'FISHM', 'TMILK', 'FRUIT', 'FDNEC', 'COFCT', 'NONAL', 
                   'ALBEV', 'TBCCO', 'FHOME', 'FDOUT', 'FOOD', 'CLOTH', 'NDFUR', 'MEDIC', 
                   'FUEL', 'TRCOM', 'RCRTN', 'EDUC', 'OTHEX', 
                   'OTDIS', 'NFOOD', 'TOTEX', 'TOTDI', 'OTREC', 'TOREC')

col_names_2012 = c('W_OID', 'FSIZE', 'WAGES', 'NETSHARE', 'CASH_ABROAD', 'CASH_DOMESTIC', 'RENTALS_REC', 'INTEREST', 
                   'PENSION', 'DIVIDENDS', 'OTHER_SOURCE', 'NET_RECEIPT', 'REGFT', 'NET_CFG', 'NET_LPR', 'NET_FISH', 
                   'NET_FOR', 'NET_RET', 'NET_MFG', 'NET_COM', 'NET_TRANS', 'NET_MIN', 'NET_CONS', 'NET_NEC', 'EAINC', 
                   'TOINC', 'T_BREAD', 'T_MEAT', 'T_FISH', 'T_MILK', 'T_FRUIT', 'T_VEG', 'T_FOOD_NEC', 'T_COFFEE', 'T_MINERAL', 
                   'T_ALCOHOL', 'T_TOBACCO', 'T_FOOD_HOME', 'T_FOOD_OUTSIDE', 'T_FOOD', 'T_CLOTH', 'T_FURNISHING', 'T_HEALTH', 
                   'T_HOUSING_WATER', 'T_TRANSPORT', 'T_COMMUNICATION', 'T_RECREATION', 'T_EDUCATION', 'T_OTHER_EXPENDITURE', 
                   'T_OTHER_DISBURSEMENT', 'T_NFOOD', 'T_TOTEX', 'T_TOTDIS', 'T_OTHREC', 'T_TOREC')

col_names_2015 = c('W_ID', 'FSIZE', 'WAGES', 'NETSHARE', 'CASH_ABROAD', 'CASH_DOMESTIC', 'RENTALS_REC', 'INTEREST', 
                   'PENSION', 'DIVIDENDS', 'OTHER_SOURCE', 'NET_RECEIPT', 'REGFT', 'NET_CFG', 'NET_LPR', 'NET_FISH', 
                   'NET_FOR', 'NET_RET', 'NET_MFG', 'NET_COM', 'NET_TRANS', 'NET_MIN', 'NET_CONS', 'NET_NEC', 'EAINC',
                   'TOINC', 'BREAD', 'MEAT', 'FISH', 'MILK', 'FRUIT', 'VEG', 'FOOD_NEC', 'COFFEE', 'MINERAL', 
                   'ALCOHOL', 'TOBACCO', 'FOOD_HOME', 'FOOD_OUTSIDE', 'FOOD', 'CLOTH', 'FURNISHING', 'HEALTH', 
                   'HOUSING_WATER', 'TRANSPORT', 'COMMUNICATION', 'RECREATION', 'EDUCATION', 'OTHER_EXPENDITURE', 
                   'OTHER_DISBURSEMENT', 'NFOOD', 'TOTEX', 'TOTDIS', 'OTHREC', 'TOREC')

col_names_2018 = c('W_PROV', 'FSIZE', 'WAGES', 'NETSHARE', 'CASH_ABROAD', 'CASH_DOMESTIC', 'RENTALS_REC', 'INTEREST', 
                   'PENSION', 'DIVIDENDS', 'OTHER_SOURCE', 'NET_RECEIPT', 'REGFT', 'NET_CFG', 'NET_LPR', 'NET_FISH', 
                   'NET_FOR', 'NET_RET', 'NET_MFG', 'NET_COM', 'NET_TRANS', 'NET_MIN', 'NET_CONS', 'NET_NEC', 'EAINC', 
                   'TOINC', 'BREAD', 'MEAT', 'FISH', 'MILK', 'FRUIT', 'VEG', 'FOOD_NEC', 'COFFEE', 'MINERAL', 
                   'ALCOHOL', 'TOBACCO', 'FOOD_HOME', 'FOOD_OUTSIDE', 'FOOD', 'CLOTH', 'FURNISHING', 'HEALTH', 
                   'HOUSING_WATER', 'TRANSPORT', 'COMMUNICATION', 'RECREATION', 'EDUCATION', 'OTHER_EXPENDITURE', 
                   'OTHER_DISBURSEMENT', 'NFOOD', 'TOTEX', 'TOTDIS', 'OTHREC', 'TOREC')

fies_2006_selected = fies_2006 %>% select(all_of(col_names_2006))
fies_2009_selected = fies_2009 %>% select(all_of(col_names_2009))
fies_2012_selected = fies_2012 %>% select(all_of(col_names_2012))
fies_2015_selected = fies_2015 %>% select(all_of(col_names_2015))
fies_2018_selected = fies_2018 %>% select(all_of(col_names_2018))
```

Extract the province number using the ID
```{r}
fies_2006_selected = fies_2006_selected %>% mutate(W_PROV = W_HCN %/% 10^16)
fies_2009_selected = fies_2009_selected %>% mutate(W_PROV = W_ID %/% 10^8)
fies_2012_selected = fies_2012_selected %>% mutate(W_PROV = W_OID %/% 10^8)
fies_2015_selected = fies_2015_selected %>% mutate(W_PROV = W_ID %/% 10^8)
```

Group by province and take the mean per column
```{r}
province_means = function(df_selected) {
  df_aggregated = df_selected %>% 
    group_by(W_PROV) %>%
    summarize(across(everything(), \(x) mean(x, na.rm = TRUE)))
  return(df_aggregated)
}

fies_2006_aggregated = province_means(fies_2006_selected)
fies_2009_aggregated = province_means(fies_2009_selected)
fies_2012_aggregated = province_means(fies_2012_selected)
fies_2015_aggregated = province_means(fies_2015_selected)
fies_2018_aggregated = province_means(fies_2018_selected)
```

Remove helper ID variable and provinces not in all FIES datasets
```{r}
# remove the helper ID variable
fies_2006_aggregated = fies_2006_aggregated %>% select(-W_HCN)
fies_2009_aggregated = fies_2009_aggregated %>% select(-W_ID)
fies_2012_aggregated = fies_2012_aggregated %>% select(-W_OID)
fies_2015_aggregated = fies_2015_aggregated %>% select(-W_ID)

# remove provinces that are in 2018 but not in the rest of the datasets
fies_2018_aggregated = fies_2018_aggregated %>% filter(W_PROV != 85 & W_PROV != 86)
```

Create new columns
```{r}
# 2006 and 2009 have a combined column for fruits and vegetables, 
# and another combined column for transportation and communication
# while 2012, 2015, and 2018 have two separate columns
fies_2012_aggregated = fies_2012_aggregated %>% 
  mutate(FRUITVEG = T_FRUIT + T_VEG) %>% 
  mutate(TRCOM = T_TRANSPORT + T_COMMUNICATION) %>%
  select(-T_FRUIT, -T_VEG, -T_TRANSPORT, -T_COMMUNICATION)
fies_2015_aggregated = fies_2015_aggregated %>% 
  mutate(FRUITVEG = FRUIT + VEG) %>%
  mutate(TRCOM = TRANSPORT + COMMUNICATION) %>%
  select(-FRUIT, -VEG, -TRANSPORT, -COMMUNICATION)
fies_2018_aggregated = fies_2018_aggregated %>% 
  mutate(FRUITVEG = FRUIT + VEG) %>%
  mutate(TRCOM = TRANSPORT + COMMUNICATION) %>%
  select(-FRUIT, -VEG, -TRANSPORT, -COMMUNICATION)
```

Add column that signifies the year for the unique idenitifier later
```{r}
fies_2006_aggregated = fies_2006_aggregated %>% mutate(YEAR = 2006)
fies_2009_aggregated = fies_2009_aggregated %>% mutate(YEAR = 2009)
fies_2012_aggregated = fies_2012_aggregated %>% mutate(YEAR = 2012)
fies_2015_aggregated = fies_2015_aggregated %>% mutate(YEAR = 2015)
fies_2018_aggregated = fies_2018_aggregated %>% mutate(YEAR = 2018)
```

Make all five datasets have the same column names (follow the 2018 column names)
```{r}
colnames(fies_2006_aggregated) = colnames(fies_2018_aggregated)
colnames(fies_2009_aggregated) = colnames(fies_2018_aggregated)
colnames(fies_2012_aggregated) = colnames(fies_2018_aggregated)
colnames(fies_2015_aggregated) = colnames(fies_2018_aggregated)
```

Replace the provincial codes with the actual province names 
```{r}
# AI was used to convert the information in the metadata into this list 
replacement_dict = list(
  "1" = "ABRA",
  "2" = "AGUSAN DEL NORTE", # manual correction made
  "3" = "AGUSAN DEL SUR",
  "4" = "AKLAN",
  "5" = "ALBAY",
  "6" = "ANTIQUE",
  "7" = "BASILAN",
  "8" = "BATAAN",
  "9" = "BATANES",
  "10" = "BATANGAS",
  "11" = "BENGUET",
  "12" = "BOHOL",
  "13" = "BUKIDNON",
  "14" = "BULACAN",
  "15" = "CAGAYAN",
  "16" = "CAMARINES NORTE",
  "17" = "CAMARINES SUR",
  "18" = "CAMIGUIN",
  "19" = "CAPIZ",
  "20" = "CATANDUANES",
  "21" = "CAVITE",
  "22" = "CEBU",
  "23" = "DAVAO",
  "24" = "DAVAO DEL SUR", # manual correction made
  "25" = "DAVAO ORIENTAL",
  "26" = "EASTERN SAMAR",
  "27" = "IFUGAO",
  "28" = "ILOCOS NORTE",
  "29" = "ILOCOS SUR",
  "30" = "ILOILO",
  "31" = "ISABELA",
  "32" = "KALINGA",
  "33" = "LA UNION",
  "34" = "LAGUNA",
  "35" = "LANAO DEL NORTE",
  "36" = "LANAO DEL SUR",
  "37" = "LEYTE",
  "38" = "MAGUINDANAO",
  "39" = "NCR, CITY OF MANILA, FIRST DISTRICT", # manual correction made
  "40" = "MARINDUQUE",
  "41" = "MASBATE",
  "42" = "MISAMIS OCCIDENTAL",
  "43" = "MISAMIS ORIENTAL",
  "44" = "MOUNTAIN PROVINCE",
  "45" = "NEGROS OCCIDENTAL",
  "46" = "NEGROS ORIENTAL",
  "47" = "COTABATO",
  "48" = "NORTHERN SAMAR",
  "49" = "NUEVA ECIJA",
  "50" = "NUEVA VIZCAYA",
  "51" = "OCCIDENTAL MINDORO",
  "52" = "ORIENTAL MINDORO",
  "53" = "PALAWAN",
  "54" = "PAMPANGA",
  "55" = "PANGASINAN",
  "56" = "QUEZON",
  "57" = "QUIRINO",
  "58" = "RIZAL",
  "59" = "ROMBLON",
  "60" = "SAMAR", # manual correction made
  "61" = "SIQUIJOR",
  "62" = "SORSOGON",
  "63" = "SOUTH COTABATO",
  "64" = "SOUTHERN LEYTE",
  "65" = "SULTAN KUDARAT",
  "66" = "SULU",
  "67" = "SURIGAO DEL NORTE",
  "68" = "SURIGAO DEL SUR",
  "69" = "TARLAC",
  "70" = "TAWI-TAWI",
  "71" = "ZAMBALES",
  "72" = "ZAMBOANGA DEL NORTE",
  "73" = "ZAMBOANGA DEL SUR",
  "74" = "NCR, SECOND DISTRICT", # manual correction made
  "75" = "NCR, THIRD DISTRICT", # manual correction made
  "76" = "NCR, FOURTH DISTRICT", # manual correction made
  "77" = "AURORA",
  "78" = "BILIRAN",
  "79" = "GUIMARAS",
  "80" = "SARANGANI",
  "81" = "APAYAO",
  "82" = "COMPOSTELA VALLEY",
  "83" = "ZAMBOANGA SIBUGAY"
)
```

Remove provinces that are only in one dataset; create unique identifier with province and year
```{r}
# Davao Del Norte, Davao Occidental, and Dinagat Islands are in HHI but not in FIES
hhi_map = hhi_map %>% 
  arrange(X) %>% 
  filter(X != "DAVAO DEL NORTE" & X != "DAVAO OCCIDENTAL" & X != "DINAGAT ISLANDS")

# create a unique identifier for each row in the FIES datasets
add_prov_year_identifier = function(df_aggregated) {
  df_aggregated = df_aggregated %>% 
    
    # 23 (Davao), 97 (Isabela City), and 98 (Cotabato City) are in FIES but not in HHI
    filter(W_PROV != 23 & W_PROV != 97 & W_PROV != 98) %>%
    
    # apply the replacement dictionary
    mutate(W_PROV = recode(W_PROV, !!!replacement_dict)) %>%
    
    # make the W_PROV column contain both the province and the year so that there is only one identifier
    mutate(W_PROV = paste(W_PROV, YEAR, sep = "_")) %>% 
    
    # remove the year
    select(-YEAR) %>%
    
    # rename W_PROV to PROV_YEAR
    rename(PROV_YEAR = W_PROV) %>%
    
    # arrange PROV_YEAR
    arrange(PROV_YEAR)
  
  return(df_aggregated)
}

fies_2006_aggregated = add_prov_year_identifier(fies_2006_aggregated)
fies_2009_aggregated = add_prov_year_identifier(fies_2009_aggregated)
fies_2012_aggregated = add_prov_year_identifier(fies_2012_aggregated)
fies_2015_aggregated = add_prov_year_identifier(fies_2015_aggregated)
fies_2018_aggregated = add_prov_year_identifier(fies_2018_aggregated)
```

Add the HHI data from two years before to the FIES data (e.g. the HHI 2004 data is added to the FIES 2006)
```{r}
fies_hhi_2006_aggregated = fies_2006_aggregated %>% mutate(HHI = hhi_map$X2004)
fies_hhi_2009_aggregated = fies_2009_aggregated %>% mutate(HHI = hhi_map$X2007)
fies_hhi_2012_aggregated = fies_2012_aggregated %>% mutate(HHI = hhi_map$X2010)
fies_hhi_2015_aggregated = fies_2015_aggregated %>% mutate(HHI = hhi_map$X2013)
fies_hhi_2018_aggregated = fies_2018_aggregated %>% mutate(HHI = hhi_map$X2016)
```

Row-bind all five datasets
```{r}
fies_hhi_all_years = rbind(fies_hhi_2006_aggregated, fies_hhi_2009_aggregated, fies_hhi_2012_aggregated,
                           fies_hhi_2015_aggregated, fies_hhi_2018_aggregated)

View(fies_hhi_all_years)
```

Export to CSV
```{r}
write.csv(fies_hhi_all_years, "fies_hhi_all_years.csv", row.names = FALSE)
```
