==> RMD Files/6 Spatial Regression.Rmd <==
---
title: "HHI_SpatialReg"
output:
  pdf_document: default
  html_document: default
date: "2024-12-08"
---
[0]: Load libraries.

```{r}
library(fuzzyjoin)
library(sf)
library(dplyr)
library(stringdist)

#Library for Spatial Regression
library(spdep)
library(spatialreg)
library(leaps)

#Libraries for Spatial Data
library(tidyverse)
library(colorspace)
library(rnaturalearth)
library(sf)
library(devtools)
library(rnaturalearthhires)
library(flextable)
library(stargazer)
```

[1]: Read the shp file.

```{r}
ph <- st_read("datasets/hhi_map/hhi_map.shp")
ph <- st_as_sf(ph)

#Rename the column in shp file
ph <- rename(ph, Province = X)
head(ph)

colnames(data)
```

[2]: Read the FIES file
```{r}
data <- read.csv("datasets/fies_hhi_all_years.csv")
head(data)

#Consider the variables which are significant from OLS regression
fies_subs <- data %>% 
  select(PROV_YEAR, NETSHARE, PENSION, REGFT, NET_LPR, NET_FISH, NET_FOR, NET_RET, NET_COM, NET_CONS, NET_NEC, HHI)
head(fies_subs, 5)

#Do log
fies_subs <- fies_subs %>%
  mutate_at(vars(NETSHARE, PENSION, REGFT, NET_LPR, NET_FISH, NET_FOR, NET_RET, NET_COM, NET_CONS, NET_NEC, HHI), funs(log(. + 1)))

fies_subs[is.na(fies_subs)] <- 0

#Extract province name from the subset
fies_subs$Province <- sapply(strsplit(fies_subs$PROV_YEAR, "_"), `[`, 1)
```
[2.5]: Do matching
```{r}
fies_aggregated <- fies_subs %>%
  group_by(Province) %>%
  summarize(
    AVG_NETSHARE = mean(NETSHARE),
    AVG_PENSION = mean(PENSION),
    AVG_REGFT = mean(REGFT),
    AVG_NET_LPR = mean(NET_LPR),
    AVG_NET_FISH = mean(NET_FISH),
    AVG_NET_FOR = mean(NET_FOR),
    AVG_NET_RET = mean(NET_RET),
    AVG_NET_COM = mean(NET_COM),
    AVG_NET_CONS = mean(NET_CONS),
    AVG_NET_NEC = mean(NET_NEC),
    AVG_HHI = mean(HHI)
  )
head(fies_aggregated, 8)

#Identify common provinces

# Identify provinces present in both datasets
common_provinces <- intersect(fies_aggregated$Province, ph$Province)

# Filter both datasets to retain only common provinces
FIES_HHI <- fies_aggregated %>%
  filter(Province %in% common_provinces)

SHP_HHI <- ph %>%
  filter(Province %in% common_provinces) %>%
  arrange(Province)

# Check dimensions of the filtered datasets
dim(FIES_HHI)
dim(SHP_HHI)
```

[3]: Match the fies data with the shp file
```{r}
FIES_HHI_SHP <- left_join(FIES_HHI, SHP_HHI, by = "Province")
colnames(FIES_HHI_SHP)
head(FIES_HHI_SHP)

head(FIES_HHI)
summary(FIES_HHI)
colnames(FIES_HHI)
```
[4]: Perform OLS Regression
```{r}
OLS_model = lm(AVG_HHI~. -Province, data = FIES_HHI)
summary(OLS_model)
```

[5]: **Method 1: Contiguity Approach**

```{r}

head(ph)
# create list of neighbors, based on contiguity, from the Philippine spatial data
PHneighbor <- poly2nb(SHP_HHI) 

# create list of centroid coordinates from the Philippine spatial data
geometry <- st_geometry(SHP_HHI)
PHcoords <- st_centroid(geometry) 

# visualize the connections between Philippine provinces
plot(PHneighbor, PHcoords) 

#Create contiguity
PH_list <- nb2listw(PHneighbor, zero.policy = TRUE) 

#Do spatial autocorrelation 
moran.test(FIES_HHI$AVG_HHI, PH_list)
moran.plot(FIES_HHI$AVG_HHI, PH_list)

# Lagrange multiplier test for spatial lag and spatial error dependencies
lm.LMtests(OLS_model, PH_list, test=c("LMlag", "LMerr"))
```
Under contiguity, the error model is significant.

```{r}
# Spatial error model
# Example of creating the lagged variable using the spatial weights matrix

spatial_error <- errorsarlm(AVG_HHI ~. -Province, data = FIES_HHI, listw = PH_list, zero.policy= TRUE)
summary(spatial_error)


stargazer(OLS_model, spatial_error, type = "text",
                    title="Regression Results from OLS and with Spatial Error")
```

[6]: **Method 2: Distance Matrix**

```{r}
lat_long = st_drop_geometry(SHP_HHI)[,c("lat", "lon")]
lat_long = as.matrix(lat_long)
PHneighbor_D = dnearneigh(lat_long, d1=0, d2=10)
PHlistw_D = nb2listw(PHneighbor_D, style="W")

# test for spatial dependence using Moran's I test
moran.test(FIES_HHI$AVG_HHI, PHlistw_D)
moran.plot(FIES_HHI$AVG_HHI, PHlistw_D)

# Lagrange multiplier test for spatial lag and spatial error dependencies
lm.LMtests(OLS_model, PHlistw_D, test=c("LMlag", "LMerr"))
```
Using distance matrix is not significant. We should stick to contiguity and error only.
